+++ src/main/java/org/w3/_2001/xmlschema/Adapter3.java
@@ -4,19 +4,19 @@ package org.w3._2001.xmlschema;
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 
 public class Adapter3
-    extends XmlAdapter<String, Integer>
+    extends XmlAdapter<String, Long>
 {
 
 
-    public Integer unmarshal(String value) {
-        return ((int)javax.xml.bind.DatatypeConverter.parseLong(value));
+    public Long unmarshal(String value) {
+        return ((long)javax.xml.bind.DatatypeConverter.parseLong(value));
     }
 
-    public String marshal(Integer value) {
+    public String marshal(Long value) {
         if (value == null) {
             return null;
         }
-        return (javax.xml.bind.DatatypeConverter.printLong((long)(int)value));
+        return (javax.xml.bind.DatatypeConverter.printLong((long)(long)value));
     }
 
 }
